Add removeMenuItems()
Added functionality to remove menu items through Menu::removeMenuItems(). Also added function Menu::reformatArray() to rearrange the order of the textObjs array after every deletion and "fill in any empty spaces."

Also made it so that the outlines of empty, dynamic menus are not drawn to the window in draw().