Add more functions and protections to Menu class
Adds additional functionality to the Menu class, such as the ability
to look up an item in the menu by its string content, and the ability
to remove the first menu item. The setBackgroundColor function now
works and every frame the background of the menu is drawn to the screen

Additional protections were added to keep the user from directly
accessing the textObjs array or changing the menu type or text origin
position when not appropriate. setTextOriginPointPriv() now handles all
the actual operations and setTextOriginPoint() acts as an in-between
for checking if calling setTextOriginPointPriv() is appropriate.
To bypass this validation, some member functions just call
setTextOriginPointPriv() directly.